star ok

+++ database/migrations/2021_01_04_150916_create_stars_table.php
@@ -21,7 +21,7 @@ class CreateStarsTable extends Migration
             $table->foreignId('user_id')
                 ->constrained()
                 ->onDelete('cascade');
-            $table->integer('point');
+            $table->float('point');
             $table->timestamps();
         });
     }
